requirements.txt
openai==1.53.1
python-dotenv==1.0.1
